+++ efinance/futures/getter.py
@@ -47,7 +47,8 @@ def get_quote_history(quote_ids: Union[str, List[str]],
                       beg: str = '19000101',
                       end: str = '20500101',
                       klt: int = 101,
-                      fqt: int = 1) -> pd.DataFrame:
+                      fqt: int = 1,
+                      **kwargs) -> pd.DataFrame:
     """
     获取期货历史行情信息
 
@@ -151,6 +152,9 @@ def get_quote_history(quote_ids: Union[str, List[str]],
                                            '名称': '期货名称'
                                            },
                                   inplace=True)
+            # NOTE 扩展接口 设定此关键词即返回 DataFrame 而不是 dict
+        if kwargs.get('return_df'):
+            df: pd.DataFrame = pd.concat(df, axis=0, ignore_index=True)
     return df